fix: MSW 오류 수정

+++ src/components/cartPage/orderModal/OrderModal.tsx
@@ -49,26 +49,28 @@ export const OrderModal = ({ closeModal }: OrderModalProps) => {
   const handleClickOrderButton = () => {
     setIsLoading(true);
 
-    order(usingPoint).then((res) => {
-      if (res.status === 409) {
-        alert(
-          '주문 도중 에러가 발생했습니다! 새로고침 후 다시 시도해보시길 바랍니다!!'
-        );
+    order(usingPoint)
+      .then((res) => {
+        if (res.status === 401) return alert('로그인이 필요합니다!');
+        if (res.status === 409 || res.status === 500)
+          return alert(
+            '주문 도중 에러가 발생했습니다! 새로고침 후 다시 시도해보시길 바랍니다!!'
+          );
+
+        const orderId = res.headers.get('Location')?.replace('/orders/', '');
 
         closeModal();
         setIsLoading(false);
-        return;
-      }
-      const orderId = res.headers.get('Location')?.replace('/orders/', '');
+        deleteAllSelectedRecoilCartItems();
+        getUserPoint().then((userPoint) => setUserPoint(userPoint.point));
+        getOrders().then((orders) => setOrders(orders));
 
-      closeModal();
-      setIsLoading(false);
-      deleteAllSelectedRecoilCartItems();
-      getUserPoint().then((userPoint) => setUserPoint(userPoint.point));
-      getOrders().then((orders) => setOrders(orders));
-
-      if (orderId) navigate('/orderDetail', { state: { orderId } });
-    });
+        if (orderId) navigate('/orderDetail', { state: { orderId } });
+      })
+      .finally(() => {
+        closeModal();
+        setIsLoading(false);
+      });
   };
 
   return (
@@ -240,7 +242,7 @@ const Style = {
     color: rgb(62, 62, 62);
   `,
   ColoredCaption: styled.span`
-    color: rgb(42, 193, 188);
+    color: rgb(24, 154, 150);
   `,
   OrderButton: styled.button`
     width: 90vw;
